+++ wp-content/themes/footmate/app/Core/Core.php
@@ -2,6 +2,12 @@
 
 namespace FM\Core;
 
+use FM\Core\Widgets;
+
 class Core
 {
+    public function __construct()
+    {
+        new Widgets();
+    }
 }